chore: move `watchDebounced` to utils/command-helpers
